# docs/HRM_swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: HRM_Project
  contact: {}
host: localhost:8081
basePath: /api
securityDefinitions:
  token:
    type: apiKey
    description: ''
    name: token
    in: header
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/roles/{id}':
    delete:
      summary: delete role
      tags:
        - Role API
      operationId: deleterole
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
    put:
      summary: update role by id
      tags:
        - Role API
      operationId: updaterolebyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/updaterolebyidrequest'
      responses:
        '200':
          description: ''
          headers: {}
...
  /roles/all:
    get:
      summary: get all role
      tags:
        - Role API
      operationId: getallrole
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: currentPage
          in: query
          required: true
          type: integer
          format: int32
          description: ''
        - name: pageSize
          in: query
          required: true
          type: integer
          format: int32
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
...
  '/rolePermissions/{id}':
    delete:
      summary: delete rolePermission
      tags:
        - RolePermission API
      operationId: deleterolePermission
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
    put:
      summary: update rolePermission by id
      tags:
        - RolePermission API
      operationId: updaterolePermissionbyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/updaterolePermissionbyidRequest'
      responses:
        '200':
          description: ''
          headers: {}
  /modules:
    post:
      summary: create new module
      tags:
        - Module API
      operationId: createnewmodule
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/createnewmodulerequest'
      responses:
        '200':
          description: ''
          headers: {}
    get:
      summary: get all module
      tags:
        - Module API
      operationId: getallmodule
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: currentPage
          in: query
          required: true
          type: integer
          format: int32
          description: ''
        - name: pageSize
          in: query
          required: true
          type: integer
          format: int32
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
  '/modules/{id}':
    delete:
      summary: delete module
      tags:
        - Module API
      operationId: deletemodule
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
    put:
      summary: update module by id
      tags:
        - Module API
      operationId: updatemodulebyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/updatemodulebyidrequest'
      responses:
        '200':
          description: ''
          headers: {}
    get:
      summary: get module by id
      tags:
        - Module API
      operationId: getmodulebyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
...
definitions:
...
  updaterolebyidrequest:
    title: updaterolebyidrequest
    example:
      roleName: demo-update
    type: object
    properties:
      roleName:
        type: string
    required:
      - roleName
...
  updaterolePermissionbyidRequest:
    title: updaterolePermissionbyidRequest
    example:
      roleId: faafd1f0-518f-4fe4-ae36-654d67da1562
      moduleId: aa221bae-884e-4d41-8ca5-f2d36beeab8c
      canRead: 1
      canWrite: 1
      canUpdate: 0
      canDelete: 0
      canApprove: 0
    type: object
    properties:
      roleId:
        type: string
      moduleId:
        type: string
      canRead:
        type: integer
        format: int32
      canWrite:
        type: integer
        format: int32
      canUpdate:
        type: integer
        format: int32
      canDelete:
        type: integer
        format: int32
      canApprove:
        type: integer
        format: int32
    required:
      - roleId
      - moduleId
      - canRead
      - canWrite
      - canUpdate
      - canDelete
      - canApprove
  createnewmodulerequest:
    title: createnewmodulerequest
    example:
      moduleName: role
    type: object
    properties:
      moduleName:
        type: string
    required:
      - moduleName
  updatemodulebyidrequest:
    title: updatemodulebyidrequest
    example:
      moduleName: role
    type: object
    properties:
      moduleName:
        type: string
    required:
      - moduleName
...
security:
  - token: []
...